+++ Relique/Relique/Views/MainWindow.MenuHandlers.cs
@@ -4,6 +4,7 @@ using Avalonia.Input;
 using Avalonia.Interactivity;
 using ItemEditor.Services;
 using Radoub.Formats.Settings;
+using Radoub.UI.Controls;
 using System;
 using System.Diagnostics;
 using System.IO;
@@ -229,9 +230,66 @@ public partial class MainWindow
                         OnSaveClick(sender, e);
                     e.Handled = true;
                     break;
+                case Key.F:
+                    if (HasFile)
+                    {
+                        OnFindClick(sender, e);
+                        e.Handled = true;
+                    }
+                    break;
+                case Key.H:
+                    if (HasFile)
+                    {
+                        OnFindReplaceClick(sender, e);
+                        e.Handled = true;
+                    }
+                    break;
+            }
+        }
+        else if (e.KeyModifiers == KeyModifiers.None)
+        {
+            switch (e.Key)
+            {
+                case Key.F3:
+                    this.FindControl<SearchBar>("FileSearchBar")?.FindNext();
+                    e.Handled = true;
+                    break;
+            }
+        }
+        else if (e.KeyModifiers == KeyModifiers.Shift)
+        {
+            switch (e.Key)
+            {
+                case Key.F3:
+                    this.FindControl<SearchBar>("FileSearchBar")?.FindPrevious();
+                    e.Handled = true;
+                    break;
             }
         }
     }
+
+    #region Search
+
+    private void OnFindClick(object? sender, RoutedEventArgs e)
+    {
+        this.FindControl<SearchBar>("FileSearchBar")?.Show(_currentFilePath);
+    }
+
+    private void OnFindReplaceClick(object? sender, RoutedEventArgs e)
+    {
+        this.FindControl<SearchBar>("FileSearchBar")?.ShowReplace(_currentFilePath);
+    }
+
+    private async void OnSearchFileModified(object? sender, EventArgs e)
+    {
+        if (!string.IsNullOrEmpty(_currentFilePath))
+        {
+            await OpenFileAsync(_currentFilePath);
+            UpdateStatus("File reloaded after replace.");
+        }
+    }
+
+    #endregion
 
     // --- Title Bar ---